requirements.txt
pytz==2022.1
google-cloud-storage==2.4.0
google-cloud-vision==2.7.3
Flask-RESTful==0.3.9
Flask-Cors==3.0.10
Flask==2.1.2
bcrypt==3.2.2
gunicorn==20.1.0